Add files via upload
Write a program in C# to calculate area of triangle, square and rectangle. Write 3 different functions for each shape to take dimensions of figure and display the area. You may create menus.
Write a console application in C# to explore different ways in which array is declared and initialized and explore different properties and methods of Array class.
Write a program in C# to read n number of values in an array and display it in reverse order